tests: first passe reorg
