--- tests/tests_unit/test_approval_modules_snapshots_clean/cdf_idp_default.yaml
@@ -1,1 +1,0 @@
-{}
