+++ tests/tests_unit/test_cli/test_approval_modules_snapshots/cdf_apm_simple_data_model.yaml
