+++ tests/tests_unit/test_cli/test_approval_modules_snapshots/cdf_asset_source_model.yaml
