+++ tests/tests_unit/test_cli/test_approval_modules_snapshots/cdf_data_pipeline_asset_debug.yaml
