+++ tests/tests_unit/test_cli/test_approval_modules_snapshots/cdf_oid_example_data.yaml
