+++ tests/tests_unit/test_cli/test_approval_modules_snapshots/example_pump_asset_hierarchy.yaml
