+++ tests/tests_unit/test_cli/test_approval_modules_snapshots/my_example_module.yaml
