+++ tests/tests_unit/test_cli/test_approval_modules_snapshots_clean/cdf_idp_default.yaml
