+++ tests/tests_unit/test_cli/test_approval_modules_snapshots_clean/cdf_infield_second_location.yaml
@@ -1,0 +1,1 @@
+{}
