+++ tests/tests_unit/test_cli/test_approval_modules_snapshots_clean/cdf_oid_example_data.yaml
