+++ tests/tests_unit/test_cli/test_approval_modules_snapshots_clean/example_pump_data_model.yaml
@@ -1,0 +1,1 @@
+{}
