+++ tests/tests_unit/test_cli/test_approval_modules_snapshots_clean/my_example_module.yaml
