+++ tests/tests_unit/test_cli/test_behavior.py
